Tests/LibWeb/Text/expected/Editing/contenteditable-arrow-navigation.txt

Home: position=0 character="h"
Right: position=1 character="e"
Right: position=2 character="l"
Right: position=3 character="l"
Right: position=4 character="o"
Right: position=5 character=" "
Down: position=23 character="e"
Left: position=22 character="r"
Up: position=4 character="o"
Right: position=5 character=" "
Right: position=6 character="рҹ‘©рҹҸјвҖҚвқӨпёҸвҖҚрҹ‘ЁрҹҸ»"
Right: position=18 character=" "
Right: position=19 character="t"
Down: position=28 character="рҹ‘©рҹҸјвҖҚвқӨпёҸвҖҚрҹ‘ЁрҹҸ»"
Up: position=19 character="t"
Down: position=28 character="рҹ‘©рҹҸјвҖҚвқӨпёҸвҖҚрҹ‘ЁрҹҸ»"
Left: position=27 character=" "
Left: position=26 character="y"
Up: position=18 character=" "
